fix plotting of ATAC peaks with backed data (closes #21)
This actually seems to be a bug in AnnData. When subsetting a backed
AnnData object (i.e. creating a view of a backed AnnData object), the
view stores its indices. These are then used by Raw.X to return a
propertly subsetted matrix. On the other hand, when subsetting a backed
Raw object, the indices are not stored, so when accessing .raw[:, x].X a
SparseDataset is returned instead of a matrix.